+++ Oxygen/oxygenengine/source/oxygen/rendering/opengl/OpenGLRenderResources.cpp
@@ -68,7 +68,7 @@ void OpenGLRenderResources::initialize()
 		const PlaneManager& planeManager = mRenderParts.getPlaneManager();
 		for (int index = 0; index < 4; ++index)
 		{
-			mPlanePatternsTexture[index].create(BufferTexture::PixelFormat::UINT_16, 0x1000, 1, planeManager.getPlanePatternsBuffer(index));
+			mPlanePatternsTexture[index].create(BufferTexture::PixelFormat::UINT_16, 0x1000, 1, planeManager.getPlanePatternsBuffer((uint8)index));
 		}
 
 		for (int index = 0; index < 4; ++index)
@@ -227,7 +227,7 @@ void OpenGLRenderResources::refresh()
 			if (!planeManager.isPlaneUsed(index))
 				continue;
 
-			const uint16* buffer = planeManager.getPlanePatternsBuffer(index);
+			const uint16* buffer = planeManager.getPlanePatternsBuffer((uint8)index);
 			const bool hasChanged = (memcmp(mPlanePatternsData[index], buffer, numPatterns * 2) != 0);
 			if (!hasChanged)
 				continue;
